Add working mobilenet food classifcation with TFLite Model Zoo
This adds one model to the model zoo and makes several necessary changes to
support.

First, it improves the TfLiteNDArray. It can contain either a tflite Tensor or a
direct ByteBuffer. Several operators now which which one it contains. It
supports correct construction from a Buffer of data for all data types. It also
has partially implemented copying support.

It also improves the BaseImageTranslator build from arguments. It adds a
Pipeline because otherwise the pipeline would be null if nothing were added to
it. It also lets you set the batchifier from the arguments (using a new
Batchifier.fromName).

Change-Id: I0c492529394320c82dd6933ad469abf90ddcb166